requirements.txt
Django==5.0.1
mysqlclient==2.2.1
python-dotenv==1.0.1
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
google-cloud-pubsub==2.18.4
